08_board User 1:n model relation

+++ boards/views.py
@@ -15,7 +15,12 @@ def create(request):
     if request.method == 'POST':
         form = BoardForm(request.POST)
         if form.is_valid():
-            board = form.save()    
+            # board를 바로 저장하지 않고, 현재 user를 넣고 저장
+            # 실제 DB에 반영 전까지의 단계를 진행하고, 그 중간에 user 정보를
+            # request.user에서 가져와서 그 후에 저장한다.
+            board = form.save(commit=False)        
+            board.user = request.user
+            board.save()
             return redirect('boards:detail', board.pk)
     # get 요청(혹은 다른 메서드)이면 기본 폼을 생성한다.
     else:
@@ -33,22 +38,30 @@ def detail(request, board_pk):
     
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else:
+            return redirect('boards:detail', board.pk)
     else:
-        return redirect('boards:detail', board.pk)
-
+        return redirect('boards:index')
 @login_required
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        form = BoardForm(request.POST, instance=board)  #1
-        if form.is_valid():
-            board = form.save()                         #2
-            return redirect('boards:detail', board.pk)
-    #GET 요청이면 (수정하기 버튼을 눌렀을때)
+    if board.user == request.user:
+        if request.method == 'POST':
+            form = BoardForm(request.POST, instance=board)  #1
+            if form.is_valid():
+                board = form.save()                         #2
+                return redirect('boards:detail', board.pk)
+        #GET 요청이면 (수정하기 버튼을 눌렀을때)
+        else:
+            form = BoardForm(instance=board)                #3
     else:
-        form = BoardForm(instance=board)                #3
-    context = {'form':form, 'board':board}
+        return redirect('boards:index')
+    context = {
+        'form':form, 
+        'board':board
+    }
     return render(request, 'boards/form.html',context)
